Relocate savegame loader
